PythonProject/lulu_ai/requirements.txt

# requirements.txt内容：
speechrecognition==3.10.0
pyaudio==0.2.13
transformers==4.37.0
torch==2.1.0
pyttsx3==2.90
pyyaml==6.0.1